KUKA-0007_Implement_Address_related_Repositories

+++ customermanagementsvc/modules/domain/src/main/java/com/kuka/h2g/domain/address/Address.java
@@ -6,8 +6,7 @@ import lombok.Value;
 @Value
 @Builder(setterPrefix = "with")
 public class Address {
-    
-    long id;
+    int id;
     String name;
     String zipCode;
     String state;
